Rename 'atom' packages to 'atomGoal'
