+++ src/main/java/uk/co/edstow/cain/atomGoal/pairGen/Distance.java
@@ -1,7 +1,7 @@
-package uk.co.edstow.cain.atom.pairGen;
+package uk.co.edstow.cain.atomGoal.pairGen;
 
-import uk.co.edstow.cain.atom.Atom;
-import uk.co.edstow.cain.atom.AtomGoal;
+import uk.co.edstow.cain.atomGoal.Atom;
+import uk.co.edstow.cain.atomGoal.AtomGoal;
 
 @SuppressWarnings("WeakerAccess")
 public class Distance {
