+++ src/main/java/uk/co/edstow/cain/atomGoal/pairGen/SimpleTransformation.java
@@ -1,9 +1,9 @@
-package uk.co.edstow.cain.atom.pairGen;
+package uk.co.edstow.cain.atomGoal.pairGen;
 
 import uk.co.edstow.cain.RegisterAllocator;
 import uk.co.edstow.cain.Transformation;
-import uk.co.edstow.cain.atom.Atom;
-import uk.co.edstow.cain.atom.AtomGoal;
+import uk.co.edstow.cain.atomGoal.Atom;
+import uk.co.edstow.cain.atomGoal.AtomGoal;
 import uk.co.edstow.cain.util.Tuple;
 
 import java.util.*;
