+++ src/main/java/uk/co/edstow/cain/atomGoal/pairGen/V1PairGenFactory.java
@@ -1,8 +1,8 @@
-package uk.co.edstow.cain.atom.pairGen;
+package uk.co.edstow.cain.atomGoal.pairGen;
 
 
-import uk.co.edstow.cain.atom.Atom;
-import uk.co.edstow.cain.atom.AtomGoal;
+import uk.co.edstow.cain.atomGoal.Atom;
+import uk.co.edstow.cain.atomGoal.AtomGoal;
 import uk.co.edstow.cain.pairgen.Context;
 import uk.co.edstow.cain.pairgen.PairGenFactory;
 import uk.co.edstow.cain.structures.GoalBag;
