+++ src/main/java/uk/co/edstow/cain/scamp5/PatternHuristic.java
@@ -2,9 +2,9 @@ package uk.co.edstow.cain.scamp5;
 
 import uk.co.edstow.cain.pairgen.Context;
 import uk.co.edstow.cain.pairgen.CostHuristic;
-import uk.co.edstow.cain.atom.pairGen.Distance;
-import uk.co.edstow.cain.atom.Atom;
-import uk.co.edstow.cain.atom.AtomGoal;
+import uk.co.edstow.cain.atomGoal.pairGen.Distance;
+import uk.co.edstow.cain.atomGoal.Atom;
+import uk.co.edstow.cain.atomGoal.AtomGoal;
 import uk.co.edstow.cain.structures.GoalBag;
 import uk.co.edstow.cain.structures.GoalPair;
