+++ src/main/java/uk/co/edstow/cain/scamp5/ThresholdScamp5ConfigGetter.java
@@ -1,6 +1,6 @@
 package uk.co.edstow.cain.scamp5;
 
-import uk.co.edstow.cain.atom.AtomGoal;
+import uk.co.edstow.cain.atomGoal.AtomGoal;
 import uk.co.edstow.cain.pairgen.Context;
 import uk.co.edstow.cain.pairgen.CostHuristic;
 import uk.co.edstow.cain.pairgen.PairGenFactory;
