+++ src/main/java/uk/co/edstow/cain/scamp5/analogue/Scamp5AnaloguePairGenFactory.java
@@ -1,11 +1,11 @@
 package uk.co.edstow.cain.scamp5.analogue;
 
-import uk.co.edstow.cain.atom.pairGen.Distance;
-import uk.co.edstow.cain.atom.pairGen.SimpleTransformation;
+import uk.co.edstow.cain.atomGoal.pairGen.Distance;
+import uk.co.edstow.cain.atomGoal.pairGen.SimpleTransformation;
 import uk.co.edstow.cain.nonlinear.LinearPairGenFactory;
 import uk.co.edstow.cain.pairgen.*;
-import uk.co.edstow.cain.atom.Atom;
-import uk.co.edstow.cain.atom.AtomGoal;
+import uk.co.edstow.cain.atomGoal.Atom;
+import uk.co.edstow.cain.atomGoal.AtomGoal;
 import uk.co.edstow.cain.Transformation;
 import uk.co.edstow.cain.scamp5.Scamp5ConfigGetter;
 import uk.co.edstow.cain.structures.GoalBag;
