+++ src/main/java/uk/co/edstow/cain/scamp5/digital/Scamp5DigitalTransformation.java
@@ -2,10 +2,9 @@ package uk.co.edstow.cain.scamp5.digital;
 
 import uk.co.edstow.cain.RegisterAllocator;
 import uk.co.edstow.cain.Transformation;
-import uk.co.edstow.cain.atom.Atom;
-import uk.co.edstow.cain.atom.AtomGoal;
-import uk.co.edstow.cain.atom.pairGen.SimpleTransformation;
-import uk.co.edstow.cain.structures.GoalPair;
+import uk.co.edstow.cain.atomGoal.Atom;
+import uk.co.edstow.cain.atomGoal.AtomGoal;
+import uk.co.edstow.cain.atomGoal.pairGen.SimpleTransformation;
 import uk.co.edstow.cain.util.Tuple;
 
 import java.util.*;
