+++ src/main/java/uk/co/edstow/cain/scamp5/emulator/ProcessingElement.java
@@ -1,6 +1,6 @@
 package uk.co.edstow.cain.scamp5.emulator;
 
-import uk.co.edstow.cain.atom.AtomGoal;
+import uk.co.edstow.cain.atomGoal.AtomGoal;
 import uk.co.edstow.cain.structures.GoalBag;
 import uk.co.edstow.cain.util.Tuple;
