+++ src/main/java/uk/co/edstow/cain/scamp5/emulator/Scamp5Verifier.java
@@ -3,8 +3,8 @@ package uk.co.edstow.cain.scamp5.emulator;
 import uk.co.edstow.cain.RegisterAllocator;
 import uk.co.edstow.cain.ReverseSearch;
 import uk.co.edstow.cain.Verifier;
-import uk.co.edstow.cain.atom.Atom;
-import uk.co.edstow.cain.atom.AtomGoal;
+import uk.co.edstow.cain.atomGoal.Atom;
+import uk.co.edstow.cain.atomGoal.AtomGoal;
 import uk.co.edstow.cain.structures.Bounds;
 import uk.co.edstow.cain.structures.Goal;
 import uk.co.edstow.cain.structures.GoalBag;
